refactor: Booking repository port 패키지 정리
Booking/Ticket domain port가 더 이상 JPA DAO가 아니므로 legacy dao package를 repository package로 이동한다. 실행 모듈과 infra adapter import만 함께 갱신해 API 동작은 유지한다.

Constraint: Booking repository interfaces are persistence-neutral domain ports after #418.

Rejected: Keeping legacy dao package | DAO naming now misrepresents the domain port boundary.

Confidence: high

Scope-risk: narrow

Directive: Do not reintroduce domain.booking.dao for Booking/Ticket repository ports.

Tested: ./gradlew compileJava compileKotlin test transitionBoundaryTest check --no-daemon

Not-tested: Manual deployed-environment smoke test.
