+++ apis/src/test/java/com/beat/apis/booking/GuestBookingServiceConcurrencyTest.java
@@ -25,7 +25,7 @@ import com.beat.apis.booking.application.GuestBookingService;
 import com.beat.apis.booking.application.dto.GuestBookingRequest;
 import com.beat.apis.booking.application.dto.GuestBookingResponse;
 import com.beat.apis.support.AbstractIntegrationTest;
-import com.beat.domain.booking.dao.BookingRepository;
+import com.beat.domain.booking.repository.BookingRepository;
 import com.beat.domain.booking.domain.BookingStatus;
 import com.beat.domain.performance.repository.PerformanceRepository;
 import com.beat.domain.performance.domain.BankName;
