+++ infra/src/main/java/com/beat/infra/persistence/booking/repository/BookingRepositoryImpl.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Repository;
 
-import com.beat.domain.booking.dao.BookingRepository;
+import com.beat.domain.booking.repository.BookingRepository;
 import com.beat.domain.booking.domain.Booking;
 import com.beat.domain.booking.domain.BookingStatus;
 import com.beat.infra.persistence.booking.entity.BookingJpaEntity;
